# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List approved products
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: integer, minimum: 0 }
        - in: query
          name: maxPrice
          schema: { type: integer, minimum: 0 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [featured, newest, price-asc, price-desc]
      responses:
        "200":
          description: Products visible in the public catalog
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
  /products/{id}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get a product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { description: Product not found }
  /agencies/{slug}:
    get:
      operationId: getPublicAgency
      tags: [catalog]
      summary: Get an active agency store by referral slug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public agency branding
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicAgency" }
        "404": { description: Agency store not found }
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order booking without online payment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderInput" }
      responses:
        "201":
          description: Order booking created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { description: Invalid order or unavailable product }
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a signed upload URL for an image
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadRequest" }
      responses:
        "200":
          description: Signed upload URL and object path
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResponse" }
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, category, description, origin, packageSize, shelfLife, storageInstructions, imageUrls, price, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        description: { type: string }
        origin: { type: string }
        packageSize: { type: string }
        shelfLife: { type: string }
        storageInstructions: { type: string }
        imageUrls:
          type: array
          items: { type: string }
        price: { type: integer, minimum: 0 }
        salePrice: { type: ["integer", "null"], minimum: 0 }
        status: { type: string, enum: [pending, approved, rejected] }
        rejectionReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer, minimum: 1 }
    OrderInput:
      type: object
      required: [customerName, phone, address, province, district, items]
      properties:
        customerName: { type: string, minLength: 1 }
        phone: { type: string, minLength: 6 }
        address: { type: string, minLength: 1 }
        province: { type: string, minLength: 1 }
        district: { type: string, minLength: 1 }
        note: { type: ["string", "null"] }
        agencySlug: { type: ["string", "null"] }
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/OrderItemInput" }
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, lineTotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: integer }
        lineTotal: { type: integer }
    Order:
      type: object
      required: [id, code, customerName, phone, address, province, district, note, total, status, agencySlug, items, createdAt]
      properties:
        id: { type: integer }
        code: { type: string }
        customerName: { type: string }
        phone: { type: string }
        address: { type: string }
        province: { type: string }
        district: { type: string }
        note: { type: ["string", "null"] }
        total: { type: integer }
        status: { type: string, enum: [new, confirmed, processing, shipped, completed, cancelled] }
        agencySlug: { type: ["string", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        createdAt: { type: string, format: date-time }
...
    PublicAgency:
      type: object
      required: [id, slug, displayName, logoUrl, primaryColor, active]
      properties:
        id: { type: integer }
        slug: { type: string }
        displayName: { type: string }
        logoUrl: { type: ["string", "null"] }
        primaryColor: { type: string }
        active: { type: boolean }
...
    UploadRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, enum: [image/jpeg, image/png, image/webp] }
    UploadResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }