# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 6 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  province: string;
  /** @minLength 1 */
  district: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  agencySlug?: string | null;
  /** @minItems 1 */
  items: OrderItemInput[];
}