# lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */
import type { UploadRequestContentType } from './uploadRequestContentType';

export interface UploadRequest {
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: UploadRequestContentType;
}