# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Mevi Commerce Hub, an agricultural commerce and order booking platform.
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}